src/main/java/com/example/webservice/HelloController.java
package com.example.webservice;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import com.example.webservice.dto.HelloResponseDto;

@RestController
public class HelloController {
	
	@GetMapping("/hello")
	public String hello() {
		return "hello";
	}
	
	@GetMapping("/hello/dto")
	public HelloResponseDto helloDto(@RequestParam("name") String name, 
									@RequestParam("age") int age) {
		return new HelloResponseDto(name, age);
	}
}
